updated sitemap and robots
